feat: move comp to folder
