+++ client/src/pages/Signup/index.js
@@ -1,0 +1,3 @@
+import component from "./Signup";
+
+export default component;
