+++ client/src/routes.js
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import { Route, Switch, withRouter } from "react-router-dom";
 import { connect } from "react-redux";
 import { fetchUser } from "./store/utils/thunkCreators";
-import Signup from "./Signup.js";
+import Signup from "./pages/Signup";
 import Login from "./Login.js";
 import { Home, SnackbarError } from "./components";
